Add experimental proxy support via envvar HTTP_PROXY
Adds proxy support to the initial redirect check with
gorequest and the chrome browser. It also always
forwards to shitty forwarding proxy regardless of
protocol.  This change made it easier to add a
proxy to both http and https.

You can use the proxy like
`export HTTP_PROXY=http://127.1:8080;gowitness ...`